# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GrowthForge AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/content:
    get:
      operationId: listContent
      tags: [content]
      summary: List content assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentAsset"
    post:
      operationId: generateContent
      tags: [content]
      summary: Generate content assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentGenerationInput"
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentAsset"
  /projects/{id}/content/{contentId}:
    get:
      operationId: getContent
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAsset"
    delete:
      operationId: deleteContent
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Social ─────────────────────────────────────────────────────────────────
  /projects/{id}/social-posts:
    get:
      operationId: listSocialPosts
      tags: [social]
      summary: List social media posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of social posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialPost"
    post:
      operationId: generateSocialPosts
      tags: [social]
      summary: Generate social media posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialPostInput"
      responses:
        "200":
          description: Generated social posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialPost"
  /projects/{id}/content-calendar:
    get:
      operationId: getContentCalendar
      tags: [social]
      summary: Get content calendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content calendar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEntry"

  # ── Email ──────────────────────────────────────────────────────────────────
  /projects/{id}/emails:
    get:
      operationId: listEmails
      tags: [email]
      summary: List email campaigns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of email campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailCampaign"
    post:
      operationId: generateEmails
      tags: [email]
      summary: Generate email campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailCampaignInput"
      responses:
        "200":
          description: Generated email campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailCampaign"

  # ── Ads ────────────────────────────────────────────────────────────────────
  /projects/{id}/ads:
    get:
      operationId: listAds
      tags: [ads]
      summary: List ad creatives
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdCreative"
    post:
      operationId: generateAds
      tags: [ads]
      summary: Generate ad creatives
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "200":
          description: Generated ad creatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdCreative"

  # ── Videos ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ContentGenerationInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
        count:
          type: integer
        prompt:
          type: string
    ContentAsset:
      type: object
      required: [id, projectId, type, title, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        seoKeywords:
          type: ["string", "null"]
        status:
          type: string
        hookStrength:
          type: ["integer", "null"]
        conversionPotential:
          type: ["integer", "null"]
        engagementPotential:
          type: ["integer", "null"]
        viralPotential:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ── Social ───────────────────────────────────────────────────────────────
    SocialPostInput:
      type: object
      required: [platforms]
      properties:
        platforms:
          type: array
          items:
            type: string
        count:
          type: integer
        prompt:
          type: string
    SocialPost:
      type: object
      required: [id, projectId, platform, caption, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        platform:
          type: string
        caption:
          type: string
        hashtags:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
        scheduledAt:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    CalendarEntry:
      type: object
      required: [date, posts]
      properties:
        date:
          type: string
        posts:
          type: array
          items:
            $ref: "#/components/schemas/SocialPost"

    # ── Email ────────────────────────────────────────────────────────────────
    EmailCampaignInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
        subject:
          type: string
        prompt:
          type: string
    EmailCampaign:
      type: object
      required: [id, projectId, type, subject, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        subject:
          type: string
        body:
          type: ["string", "null"]
        previewText:
          type: ["string", "null"]
        status:
          type: string
        openRate:
          type: ["number", "null"]
        clickRate:
          type: ["number", "null"]
        createdAt:
          type: string

    # ── Ads ──────────────────────────────────────────────────────────────────
    AdInput:
      type: object
      required: [platform]
      properties:
        platform:
          type: string
        count:
          type: integer
        prompt:
          type: string
    AdCreative:
      type: object
      required: [id, projectId, platform, headline, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        platform:
          type: string
        headline:
          type: string
        description:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        type:
          type: string
        hookStrength:
          type: ["integer", "null"]
        conversionPotential:
          type: ["integer", "null"]
        status:
          type: string
        createdAt:
          type: string

    # ── Videos ───────────────────────────────────────────────────────────────
...