# lib/api-zod/src/generated/types/adCreative.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowthForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdCreative {
  id: number;
  projectId: number;
  platform: string;
  headline: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cta?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  type?: string;
  /** @nullable */
  hookStrength?: number | null;
  /** @nullable */
  conversionPotential?: number | null;
  status: string;
  createdAt: string;
}