# lib/api-zod/src/generated/types/calendarEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowthForge AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialPost } from './socialPost';

export interface CalendarEntry {
  date: string;
  posts: SocialPost[];
}