# lib/api-zod/src/generated/types/emailCampaignInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowthForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailCampaignInput {
  type: string;
  subject?: string;
  prompt?: string;
}